requirements.txt
pytest
requests
python-dotenv
faker
pydantic
pytest-assume
pytest-xdist
allure-pytest